+++ src/main/java/lagatrix/server/exceptions/manager/partition/PartitionException.java
@@ -1,4 +1,6 @@
-package lagatrix.server.exceptions.manager;
+package lagatrix.server.exceptions.manager.partition;
+
+import lagatrix.server.exceptions.manager.ManagerException;
 
 /**
  * This exception repersents all errors related to the partitions.