+++ src/main/java/lagatrix/server/exceptions/manager/user/UserException.java
@@ -1,0 +1,17 @@
+package lagatrix.server.exceptions.manager.user;
+
+import lagatrix.server.exceptions.manager.ManagerException;
+
+/**
+ * This exception repersents all errors related to the users.
+ *
+ * @author javier
+ * @since 1.0
+ */
+public class UserException extends ManagerException{
+    
+    public UserException(String component) {
+        super(String.format("Can't obtain user component: %s", component));
+    }
+    
+}