+++ src/main/java/lagatrix/server/manager/information/hardware/hardware_status/temperature/cpu/RaspberryPiTemperature.java
@@ -1,6 +1,6 @@
-package lagatrix.server.manager.hardware_status.temperature.cpu;
+package lagatrix.server.manager.information.hardware.hardware_status.temperature.cpu;
 
-import lagatrix.server.exceptions.manager.TemperatureException;
+import lagatrix.server.exceptions.manager.hardware.status.TemperatureException;
 import lagatrix.server.tools.command.CommandExecutor;
 
 /**