+++ src/main/java/lagatrix/server/manager/information/hardware/hardware_status/use/UseManager.java
@@ -1,7 +1,7 @@
-package lagatrix.server.manager.hardware_status.use;
+package lagatrix.server.manager.information.hardware.hardware_status.use;
 
 import lagatrix.server.exceptions.command.CommandException;
-import lagatrix.server.exceptions.manager.UseException;
+import lagatrix.server.exceptions.manager.hardware.status.UseException;
 import lagatrix.server.tools.command.CommandExecutor;
 import lagatrix.server.tools.command.CommandResponse;