+++ src/test/java/lagatrix/server/tools/temperature/TemperatureDeviceDetectorTest.java
@@ -2,8 +2,8 @@ package lagatrix.server.tools.temperature;
 
 import lagatrix.server.tools.hawrdare_status.DevicesEnum;
 import lagatrix.server.tools.hawrdare_status.TemperatureDeviceDetector;
-import lagatrix.server.manager.hardware_status.temperature.cpu.NormalCPUTemperature;
-import lagatrix.server.manager.hardware_status.temperature.cpu.RaspberryPiTemperature;
+import lagatrix.server.manager.information.hardware.hardware_status.temperature.cpu.NormalCPUTemperature;
+import lagatrix.server.manager.information.hardware.hardware_status.temperature.cpu.RaspberryPiTemperature;
 import lagatrix.server.tools.command.CommandExecutor;
 import org.junit.Test;
 import static org.junit.Assert.*;